Add service data to BLE payload
I found out the service data wasn't being included when the project ran on my device, so I created an issue at https://github.com/tinygo-org/bluetooth/issues/306. A maintainer kindly created a pull request with a fix already, so I'm updating to that specific commit of the TinyGo bluetooth library for now.